refactor: RoleEnum 패키지 변경
RoleEnum, RoleEnumConverter 패키지 변경
