+++ src/main/java/com/gg/mafia/domain/member/dao/RoleDaoCustom.java
@@ -3,7 +3,7 @@ package com.gg.mafia.domain.member.dao;
 
 import java.util.Optional;
 import com.gg.mafia.domain.member.domain.Role;
-import com.gg.mafia.domain.member.model.RoleEnum;
+import com.gg.mafia.domain.model.RoleEnum;
 
 public interface RoleDaoCustom {
     Optional<Role> findByValue(RoleEnum value);
