+++ src/main/java/com/gg/mafia/domain/member/dao/RoleDaoImpl.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import com.gg.mafia.domain.member.domain.QRole;
 import com.gg.mafia.domain.member.domain.Role;
-import com.gg.mafia.domain.member.model.RoleEnum;
+import com.gg.mafia.domain.model.RoleEnum;
 
 @RequiredArgsConstructor
 public class RoleDaoImpl implements RoleDaoCustom {
